ecourseapiv1/requirements.txt
asgiref==3.7.2
automium==0.2.6
automium-web==0.1.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Django==5.0.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
legacy==0.1.7
MarkupSafe==2.1.5
pillow==10.2.0
PyMySQL==1.1.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
Werkzeug==3.0.1
